% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Instance_Identifiers.R
\name{Add_Instance_Identifiers}
\alias{Add_Instance_Identifiers}
\title{Add Instance Identifiers}
\usage{
Add_Instance_Identifiers(instances)
}
\arguments{
\item{instances}{a named list of jaw instance data frames titled with the OU Speech Motor Lab nomenclature}
}
\value{
a parallel named list of identifiers based on the nomenclature
}
\description{
Add Instance Identifiers
}
\examples{
\dontrun{Add_Instance_Identifiers(instance_directory)}
}
