% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Instance_Repetitions.R
\name{Add_Instance_Repetitions}
\alias{Add_Instance_Repetitions}
\title{Add Instance Repetitions}
\usage{
Add_Instance_Repetitions(instances)
}
\arguments{
\item{instances}{a named list of jaw instance data frames titled with the OU Speech Motor Lab nomenclature}
}
\value{
a parallel named list of repetition number based on the nomenclature
}
\description{
Add Instance Repetitions
}
\examples{
\dontrun{Add_Instance_Repetitions(instance_directory)}
}
