% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Time_Cols.R
\name{Add_Time_Cols}
\alias{Add_Time_Cols}
\title{Add Time Columns}
\usage{
Add_Time_Cols(instances)
}
\arguments{
\item{instances}{a list of jaw utterance instance data frames}
}
\value{
a list of jaw utterance instance data frames with a time column "Time" of time since utterance onset
and a recentered time column "CenteredTime" of time since the lowest point of jaw vertical displacement,
both in milliseconds
}
\description{
Add Time Columns
}
\examples{
\dontrun{Add_Time_Cols(instances = instance_directory)}
}
