% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Velocity_Cols.R
\name{Add_Velocity_Cols}
\alias{Add_Velocity_Cols}
\title{Add Velocity Columns}
\usage{
Add_Velocity_Cols(instances)
}
\arguments{
\item{instances}{a list of jaw utterance instance data frames}
}
\value{
a list of jaw utterance instance data frames with a velocity column "Velocity"
and a rescaled column "RescaledVelocity" of [-1, 1] for layered visualization
}
\description{
Add Velocity Columns
}
\examples{
\dontrun{Add_Velocity_Cols(instance_directory)}
}
