% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Z_Cols.R
\name{Add_Z_Cols}
\alias{Add_Z_Cols}
\title{Add Z Columns}
\usage{
Add_Z_Cols(instances)
}
\arguments{
\item{instances}{a list of jaw utterance instance data frames}
}
\value{
a list of jaw utterance instance data frames with a recentered vertical position column "RecenteredZ"
and a rescaled column "RescaledZ" of [-1, 1] for layered visualization
}
\description{
Add Z Columns
}
\examples{
\dontrun{Add_Z_Cols(instances)}
}
