% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Add_Utterance_Attributes.R
\name{All_Add_Utterance_Attributes}
\alias{All_Add_Utterance_Attributes}
\title{All Add Utterance Attributes}
\usage{
All_Add_Utterance_Attributes()
}
\value{
This function does not return any value. Instead, it uploads all jaw data csv files in the given format into the parent environment as data frames.
}
\description{
This function uploads all csv's within a given file directory and format into the parent environment. It then adds a time column and other trial attributes to each trial.
}
\details{
WARNING: This function assumes all data is Jaw Motion Data within the extracted utterance format using the MATLAB Functions "Extract_Jaw_Interval_Data" and "All_Jaw_Utterance_To_CSV". Do not use other data types.

\itemize{
  \item \code{Time}: A new column added to represent the time since onset of each row, counted in milliseconds.
  \item \code{Subject}: The subject number based on the file name.
  \item \code{Trial}: The trial number based on the file name.
  \item \code{Rate}: A variable that designates the rate of speaking (Slow, Fast, or Other) based on the file name.
    \itemize{
      \item 'S' -> "Slow"
      \item 'F' -> "Fast"
      \item Other -> "Other"
    }
  \item \code{Stress}: A variable that designates the vowel stress condition (Stressed, Unstressed, or Other) based on the file name.
  \itemize{
      \item 'S' -> "Stressed"
      \item 'U' -> "Unstressed"
    }
  \item \code{Vowel}: A variable of the vowel spoken based on the file name.
  \item \code{Consonant}: A variable of the consonant spoken based on the file name.
  \item \code{Repetition}: The repetition number from the file name.
  \item \code{Hearing}: A variable that designates the hearing type based on the file name.
  \item \code{Age}: A variable that identifies the age group (Adult or Child) based on the file name.
   \itemize{
      \item 'A' -> "Adult"
      \item 'C' -> "Child"
    }
  \item \code{Fix}: Currently Unknown to be fixed
}

A warning will be issued if a file cannot be read or if a column is missing within the data.
}
\examples{
\dontrun{All_Add_Utterance_Attributes()}
}
