% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Instance_Label.R
\name{Create_Instance_Identifier}
\alias{Create_Instance_Identifier}
\title{Create Instance Indentifier}
\usage{
Create_Instance_Identifier(
  Subject,
  Rate,
  Stress,
  Vowel,
  Consonant,
  Hearing,
  Age
)
}
\arguments{
\item{Subject}{an integer representing an existing subject number: "18" or "19"}

\item{Rate}{a string of the desired utterance rate: "Fast" or "Slow"}

\item{Stress}{a string of the desired vowel stress type: "Stressed" or "Unstressed"}

\item{Vowel}{a 1-character string of the desired vowel spoken: 'a' or 'e'}

\item{Consonant}{a 1-character string of the desired consonant spoken: 'd' or 't'}

\item{Hearing}{an all capital string of the desired hearing type: "NH"}

\item{Age}{a string of the desired age category: "Adult" or "Child"}
}
\value{
the instance identifier for the desired utterance type
}
\description{
Create Instance Indentifier
}
\examples{
{Create_Instance_Identifier(Subject = 19,
Rate = "Fast", Stress = "Stressed", Vowel = "a",
Consonant = "d", Hearing = "NH", Age = "Adult")}
}
