% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Instance_Plot.R
\name{Create_Instance_Plot}
\alias{Create_Instance_Plot}
\title{Create_Instance_Plot}
\usage{
Create_Instance_Plot(
  X,
  Y,
  Subject,
  Rate,
  Stress,
  Vowel,
  Consonant,
  Hearing,
  Age,
  Instances,
  IDs
)
}
\arguments{
\item{X}{a vector of instance data to display}

\item{Y}{a vector of instance data to display}

\item{Subject}{an integer representing an existing subject number: "18" or "19"}

\item{Rate}{a string of the desired utterance rate: "Fast" or "Slow"}

\item{Stress}{a string of the desired vowel stress type: "Stressed" or "Unstressed"}

\item{Vowel}{a 1-character string of the desired vowel spoken: 'a' or 'e'}

\item{Consonant}{a 1-character string of the desired consonant spoken: 'd' or 't'}

\item{Hearing}{an all capital string of the desired hearing type: "NH"}

\item{Age}{a string of the desired age category: "Adult" or "Child"}

\item{Instances}{a named list of jaw instance data frames titled with the OU Speech Motor Lab nomenclature}

\item{IDs}{a parallel named list of jaw instance IDs titled with the OU Speech Motor Lab nomenclature}
}
\value{
a plot of the selected instance data
}
\description{
Create_Instance_Plot
}
\examples{
\dontrun{Create_Instance_Plot(
X = "Time",
Y = "Z",
Subject = "18",
Rate = "Fast",
Stress = "Stressed",
Vowel = 'a',
Consonant = 'd',
Hearing = "NH",
Age = "Adult",
Instances = instance_directory,
IDs = id_directory)}
}
