% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upload_Jaw_Data.R
\name{Upload_Jaw_Data}
\alias{Upload_Jaw_Data}
\title{Upload Jaw Data}
\usage{
Upload_Jaw_Data(dir, list)
}
\arguments{
\item{dir}{a directory/file path containing only csvs of jaw motion data within the given file format; if no directory is given, the function will use the current working directory}

\item{list}{a list of jaw instance data frames to add to}
}
\value{
a named list of jaw instance data frames with added data from files in the directory
}
\description{
This function uploads all csv's within a given file directory and format into a named list.
}
\details{
WARNING: This function assumes all data is Jaw Motion Data within the extracted utterance format using the MATLAB Functions "Extract_Jaw_Interval_Data" and "All_Jaw_Utterance_To_CSV". Do not use other data types.
}
\examples{
\dontrun{Upload_Jaw_Data()}
\dontrun{Upload_Jaw_Data("C://Users//JohnDoe//
OneDrive//Jaw Motion Research//Data//Jaw-Motion-CSV-ALL")}
}
